Force float32.

+++ calc_xi.py
@@ -24,14 +24,14 @@ def calc_xi(dat,ran,bins=None):
     ddc = dat['DEC'].astype('float32')
     dcz = dat['CHI'].astype('float32')
     if not 'WT' in dat.keys():
-        dwt = np.ones_like(dat['RA'])
+        dwt = np.ones_like(dat['RA']).astype('float32')
     else:
         dwt = dat['WT'].astype('float32')
     rra = ran['RA' ].astype('float32')
     rdc = ran['DEC'].astype('float32')
     rcz = ran['CHI'].astype('float32')
     if not 'WT' in ran.keys():
-        rwt = np.ones_like(ran['RA'])
+        rwt = np.ones_like(ran['RA']).astype('float32')
     else:
         rwt = ran['WT'].astype('float32')
     # Bin edges are specified in Mpc/h, if nothing
